renamed Digits to Converter, renamed PermissiveFormatter to FallbackFormatter, updated README
